requirements.txt
blinker==1.9.0
cffi==2.0.0
click==8.3.1
cryptography==46.0.5
Flask==3.1.3
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pycparser==3.0
PyMySQL==1.1.2
python-dotenv==1.2.1
SQLAlchemy==2.0.47
typing_extensions==4.15.0
Werkzeug==3.1.6
pytest
pytest-cov
pytest-mock
